+++ apps/router-e2e/index.js
@@ -1,1 +1,5 @@
-import 'expo-router/entry';
+if (process.env.EXPO_PUBLIC_USE_RSC) {
+  require('expo-router/entry-rsc');
+} else {
+  require('expo-router/entry');
+}